Added ability to impute with a float
I’ve added the ability to impute values with a float, so that I can put in limiting magnitudes. Also added a detection flag column to the mags dataset. I haven’t added any checking to see if the limit is being imputed into the column for alpha, but it’ll have to do for now.